server/scripts/requirements.txt
anthropic
python-dotenv
argparse
logging
yfinance
requests
GoogleNews
concurrent.futures
hashlib
